More strategies for XAGs. (#20)
* pebbling view

* pebbling strategy with template solver.

* fixing max step in pebbling strategy stats.

* fix test for pebbling strat. with z3solver.

* z3 solver without limits on num. pebbles.

* weighted pebbling strategy.

* XAG with multi-input xor nodes.

* in-place solver.

* in-place solver and fix warnings.

* comments for in-place encoding.

* fixing duplicate symbols.

* fixing comp errors.

* fixing XAG strategy.

* XAG pebbling strategy and test.

* fixing dep on Z3.

* update Travis.

* fixing comp.

* fix comp.

* fixing steps sorting.

* show progress.

* update lorina.

* consider dependent outputs.

* test case 11.

* create experiments and update libs.

* solver manager.

* timeout.

* experiments stats methods.

* XAG min depth strategy and tests.

* fixes.

* introduce multiplicative depth.

* XAG stats.

* weighted pebbling XAG.

* small fixes.

* update Travis.

* fixing compilation.

* fix xor outputs.

* update libs.

* fit_depth XAG mapping strategy.

* update mockturtle.

* warnings.

* new low depth algorithm.

* parity function with one input.

* sorting network for weight constraint.

* conflict limit.

* optimize pebbling solution.

* reduce box nodes inputs.

* enable both decrement_pebbles and optimize.

* optimize weight.

* solver set options.

* small fix.

* small changes.

* optimize weight.

* reformatting.

* compute and uncompute levels actions fixed.

* renaming and pebbling strategy optimization.

* igraph C library integrated.

* cleaning-up tests for xag strategies.

* buffer action.

* fixing buffer bug.

* small fix.

* fixing model clean-up issue.

* small fix.

* new test case.

* tests for fast low T-count XAG synthesis.

* level by level low T-count XAG synthesis.

* decomposition of large compatibility graphs.

* update gitignore.

* virtual synthesis of XAGs.

* release ancillae used for copies.

* fix seg fault in optimization.

* compute qubits offset inside tracer.

* alap method to get XAG levels.

* low T-count implementation with abstract XAG.

* low depth synthesis of abstract XAGs.

* fix alap abstract XAG strategy.

* remove igraph.

* use dynamic bitset from easy.

* clean up unused files.

* fix compilation with gcc.

* fix compilation gcc.

+++ lib/mockturtle/mockturtle/generators/random_logic_generator.hpp
@@ -1,5 +1,5 @@
 /* mockturtle: C++ logic network library
- * Copyright (C) 2018  EPFL
+ * Copyright (C) 2018-2019  EPFL
  *
  * Permission is hereby granted, free of charge, to any person
  * obtaining a copy of this software and associated documentation
@@ -90,15 +90,15 @@ public:
     }
 
     /* generate gates */
-    std::mt19937 rng( seed );
-    std::uniform_int_distribution<int> rule_dist( 0, _gens.size()-1u );
+    std::mt19937 rng( static_cast<unsigned int>( seed ) );
+    std::uniform_int_distribution<int> rule_dist( 0, static_cast<int>( _gens.size() - 1u ) );
 
     auto gate_counter = ntk.num_gates();
     while ( gate_counter < num_gates )
     {
       auto const r = _gens.at( rule_dist( rng ) );
 
-      std::uniform_int_distribution<int> dist( 0, fs.size()-1 );
+      std::uniform_int_distribution<int> dist( 0, static_cast<int>( fs.size() - 1 ) );
       std::vector<signal> args;
       for ( auto i = 0u; i < r.num_args; ++i )
       {
